pyproject.toml
# pyproject.toml
[project]
name = "dc-etl"
dynamic = ["version"]
authors = [
    { name = "Chris Rossi", email = "chris@christophermrossi.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering"
]
description = "Tools for loading climate data into IPFS clusters"
readme = "README.md"
keywords = ["Climate", "Zarr", "ETL", "IPFS"]
requires-python = ">=3.11"
dependencies = [
    "fsspec",
    "kerchunk[hdf]",
    "ipldstore @ git+https://github.com/dClimate/ipldstore@v2.1.3",
    "numcodecs",
    "numpy",
    "orjson",
    "pyyaml",
    "xarray[io]",
]

[project.urls]
Homepage = "https://www.dclimate.net/"

[project.entry-points.fetcher]
cpc = "dc_etl.fetchers.cpc:CPCFetcher"
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.extractor]
netcdf = "dc_etl.extractors.netcdf:NetCDFExtractor"
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.combiner]
default = "dc_etl.combine:DefaultCombiner"
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.combine_preprocessor]
fix_fill_value = "dc_etl.combine_preprocessors:fix_fill_value"
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.combine_postprocessor]
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.transformer]
composite = "dc_etl.transformers:Composite"
rename_dims = "dc_etl.transformers:rename_dims"
normalize_longitudes = "dc_etl.transformers:normalize_longitudes"
compress = "dc_etl.transformers:compress"
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.loader]
ipld = "dc_etl.ipld.loader:IPLDLoader"
testing = "tests.unit.conftest:mock_entry_point"

[project.entry-points.ipld_publisher]
local_file = "dc_etl.ipld.local_file:LocalFileIPLDPublisher"
testing = "tests.unit.conftest:mock_entry_point"

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
]
dev = [
    "nox",
    "ruff"
]
doc = [
    "sphinx",
]
examples = [
    "docopt",
    "python-dateutil",
]
all = [
    "dc-etl[dev,testing,doc,examples]",
]

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "no-local-version"
write_to = "dc_etl/_version.py"

[tool.black]
line-length = 119

[tool.flake8]
# flake8 and black disagree on E203, E701, and W503
ignore = ["E203", "E701", "W503"]
max-line-length = 119
